HBASE-29558: Addendum to fix PR validation warnings on the lines of #7276
Also includes the following additional changes done in PR #7274 from branch-2:
- Reverted the change admin test, adding back the dedicated driver and renaming the ruby test file.
- Removed BaseTestShell as it is no longer needed after serving the requirement of admin test via its own driver
- The base setUp() method now uses a single node cluster and only admin and RS group tests override it to 3 nodes.
