+++ hbase-shell/src/test/java/org/apache/hadoop/hbase/client/TestQuotasShell.java
@@ -24,7 +24,7 @@ import org.junit.ClassRule;
 import org.junit.experimental.categories.Category;
 
 @Category({ ClientTests.class, LargeTests.class })
-public class TestQuotasShell extends BaseTestShell {
+public class TestQuotasShell extends AbstractTestShell {
   @ClassRule
   public static final HBaseClassTestRule CLASS_RULE =
     HBaseClassTestRule.forClass(TestQuotasShell.class);
