+++ hbase-shell/src/test/ruby/hbase/admin_test.rb
